5. AS5048 SPI 코드 작성

+++ Core/Inc/vardef.h
@@ -1,12 +1,12 @@
 
 typedef unsigned char uint8;
 
-typedef unsigned int uint16;
+typedef unsigned short uint16;
 
 typedef unsigned long uint32;
 
 typedef signed char sint8;
 
-typedef signed int sint16;
+typedef signed short sint16;
 
 typedef signed long sint32;
